+++ packages/@ui/src/feedback/export-modal-light/index.ts
@@ -1,0 +1,6 @@
+import type { TableColumnInfo, ExportModalLightOptions, ExportModalLightOpenParams } from './types'
+import SunnyExportModalLight from './SunnyExportModalLight.vue'
+
+export { SunnyExportModalLight }
+export type { TableColumnInfo, ExportModalLightOptions, ExportModalLightOpenParams }
+export * from './use-export-modal-light'