+++ packages/@ui/src/index.ts
@@ -69,6 +69,7 @@ export * from "./composite/customize-select";
 export * from './data/resource-tree/types';
 export * from './data/resource-tree/use-sunny-resource-tree';
 export * from './feedback/export-modal';
+export * from './feedback/export-modal-light';
 export * from './feedback/import-modal';
 export * from './entry/qrcode-reader/types';
 export { SunnyQrcodeReader } from './entry/qrcode-reader';