+++ packages/@ui/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sunny-base-web/ui
 
+## 0.8.48
+
+### Patch Changes
+
+- feat: 新增 SunnyExportModalLight 轻量导出组件及 useExport hook
+  - @sunny-base-web/icons@0.8.48
+  - @sunny-base-web/locales@0.8.48
+  - @sunny-base-web/utils@0.8.48
+
+## 0.0.0-0.8.48-20260430080121
+
+### Patch Changes
+
+- feat: 新增 SunnyExportModalLight 轻量导出组件及 useExport hook
+  - @sunny-base-web/icons@0.0.0-0.8.48-20260430080121
+  - @sunny-base-web/locales@0.0.0-0.8.48-20260430080121
+  - @sunny-base-web/utils@0.0.0-0.8.48-20260430080121
+
 ## 0.8.47
 
 ### Patch Changes